+++ CHANGELOG.md
@@ -16,6 +16,7 @@ We go to the next version after each release on
 ### Fixed
 
 - Activities button error in GNOME Shell mobile.
+- `.set_icon_size` is not a function error (Date Menu).
 
 ### Added